Updated the relief tool, added a help string and removed aerial perspective since it works only so-so with high height maps

+++ soillib/io/img.hpp
@@ -16,7 +16,7 @@ struct img {
 
   img(){};
   
-  img(const size_t width, const size_t height):
+  img(const uint32_t width, const uint32_t height):
   width{width},height{height}{ 
     allocate();
   }
